// src/generated-api/src/models/DomainFavoritePublisherList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Affiliate Backend API
 * API Server for Affiliate Backend Application
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DomainFavoritePublisherListItem } from './DomainFavoritePublisherListItem';
import {
    DomainFavoritePublisherListItemFromJSON,
    DomainFavoritePublisherListItemFromJSONTyped,
    DomainFavoritePublisherListItemToJSON,
    DomainFavoritePublisherListItemToJSONTyped,
} from './DomainFavoritePublisherListItem';

/**
 * 
 * @export
 * @interface DomainFavoritePublisherList
 */
export interface DomainFavoritePublisherList {
    /**
     * 
     * @type {string}
     * @memberof DomainFavoritePublisherList
     */
    createdAt?: string;
    /**
     * 
     * @type {string}
     * @memberof DomainFavoritePublisherList
     */
    description?: string;
    /**
     * Optional: Include items when fetching with details
     * @type {Array<DomainFavoritePublisherListItem>}
     * @memberof DomainFavoritePublisherList
     */
    items?: Array<DomainFavoritePublisherListItem>;
    /**
     * 
     * @type {number}
     * @memberof DomainFavoritePublisherList
     */
    listId?: number;
    /**
     * 
     * @type {string}
     * @memberof DomainFavoritePublisherList
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof DomainFavoritePublisherList
     */
    organizationId?: number;
    /**
     * 
     * @type {string}
     * @memberof DomainFavoritePublisherList
     */
    updatedAt?: string;
}

/**
 * Check if a given object implements the DomainFavoritePublisherList interface.
 */
export function instanceOfDomainFavoritePublisherList(value: object): value is DomainFavoritePublisherList {
    return true;
}

export function DomainFavoritePublisherListFromJSON(json: any): DomainFavoritePublisherList {
    return DomainFavoritePublisherListFromJSONTyped(json, false);
}

export function DomainFavoritePublisherListFromJSONTyped(json: any, ignoreDiscriminator: boolean): DomainFavoritePublisherList {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['created_at'] == null ? undefined : json['created_at'],
        'description': json['description'] == null ? undefined : json['description'],
        'items': json['items'] == null ? undefined : ((json['items'] as Array<any>).map(DomainFavoritePublisherListItemFromJSON)),
        'listId': json['list_id'] == null ? undefined : json['list_id'],
        'name': json['name'] == null ? undefined : json['name'],
        'organizationId': json['organization_id'] == null ? undefined : json['organization_id'],
        'updatedAt': json['updated_at'] == null ? undefined : json['updated_at'],
    };
}

export function DomainFavoritePublisherListToJSON(json: any): DomainFavoritePublisherList {
    return DomainFavoritePublisherListToJSONTyped(json, false);
}

export function DomainFavoritePublisherListToJSONTyped(value?: DomainFavoritePublisherList | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['createdAt'],
        'description': value['description'],
        'items': value['items'] == null ? undefined : ((value['items'] as Array<any>).map(DomainFavoritePublisherListItemToJSON)),
        'list_id': value['listId'],
        'name': value['name'],
        'organization_id': value['organizationId'],
        'updated_at': value['updatedAt'],
    };
}